Move function to shared helper file
